Better separate application and library code
Move all application code into a new top level "application" directory,
and put the "resources" folder into that directory as well so that we
can use it to store resources Qt needs (like the icon).
